Moshi

Documentation for Moshi.

Moshi.Data.Self

Moshi.Data.convert_singleton_bottom

Moshi.Data.data_type_name

Moshi.Data.is_data_type

Moshi.Data.isa_variant

Moshi.Data.show_variant

Moshi.Data.show_variant

Moshi.Data.variant_fieldnames

Moshi.Data.variant_fieldtypes

Moshi.Data.variant_kind

Moshi.Data.variant_name

Moshi.Data.variant_storage

Moshi.Data.variant_type

Moshi.Data.variants

Moshi.Data.variants

Moshi.Data.@data

Moshi.Data.@pass

Moshi.Derive.@derive

Moshi.Data.Self — Constant

Syntax placeholder for self type.

Moshi.Data.convert_singleton_bottom — Method

Convert a singleton variant with Union{} as type parameters to a matching singleton variant with the correct type parameters.

Note

Base.convert(::Type{YourData.Type}, x::YourData.Type{Union{}}) fall back to this method so it can be overloaded if necessary. This method fallback to a generated method by @data by default.

Moshi.Data.data_type_name — Method

public Moshi.Data.data_type_name(x) -> Symbol

Return the data type name of the given variant.

Moshi.Data.is_data_type — Method

public Moshi.Data.is_data_type(x) -> Bool

Check if the given object is a variant of a algebraic data type.

Moshi.Data.isa_variant — Method

public Moshi.Data.isa_variant(x, variant::Type) -> Bool

Check if the given variant is an instance of the given variant type.

Moshi.Data.show_variant — Method

public Moshi.Data.show_variant(io::IO, x)

Print the variant to the given IO stream in a single line.

Moshi.Data.show_variant — Method

public Moshi.Data.show_variant(io::IO, x)

Print the variant to the given IO stream in multiple lines.

Moshi.Data.variant_fieldnames — Method

public Moshi.Data.variant_fieldnames(value)

Return the field names of the variant.

Moshi.Data.variant_fieldtypes — Method

public Moshi.Data.variant_fieldtypes(value)

Return the field types of the variant.

Moshi.Data.variant_kind — Method

public Moshi.Data.variant_kind(x) -> VariantKind

Return the kind of the variant, can be Singleton, Anonymous, or Named.

Moshi.Data.variant_name — Method

public Moshi.Data.variant_name(x) -> Symbol

Return the name of the variant.

Moshi.Data.variant_storage — Method

public Moshi.Data.variant_storage(value)

Return the storage object of the variant.

Moshi.Data.variant_type — Method

public Moshi.Data.variant_type(x)

Return the variant type of the given variant.

Moshi.Data.variants — Method

public Moshi.Data.variants(x::Type) -> Tuple

Return a tuple of variants of the given data type.

Moshi.Data.variants — Method

public Moshi.Data.variants(x::Type) -> Tuple

Return a tuple of variants of the given data type.

Moshi.Data.@data — Macro

@data <head> <variants>

Create a new algebraic data type (also known as a sum type) with the given head and variants.

Moshi.Data.@pass — Macro

@pass [priority] <function definition>

Add a function to the list of functions to be called when emitting code.

Optionally, you can specify a priority for the function. The default priority is 5. Lower priority number functions are called first.

Moshi.Derive.@derive — Macro

@derive <type>[<traits>...]

Automatically derive traits for a concrete type. The following traits are supported:

PartialEq

Hash

Tree